Rename data_pg to data
