+++ tests/server/test_incremental_deploy.py
@@ -28,7 +28,7 @@ from inmanta.agent.agent import Agent
 from inmanta.server import SLICE_SERVER, SLICE_AGENT_MANAGER
 import uuid
 from datetime import datetime
-from inmanta import data_pg as data, const, config
+from inmanta import data, const, config
 from inmanta.const import ResourceState, ResourceAction
 import logging
 from typing import List, Dict, Any
