+++ tests/test_agent_manager.py
@@ -21,7 +21,7 @@ import asyncio
 
 import pytest
 from inmanta.server.agentmanager import AgentManager
-from inmanta import data_pg as data
+from inmanta import data
 from inmanta.protocol import Result
 from utils import assert_equal_ish, UNKWN
